backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.127.0",
    "google-adk>=1.18.0",
    "google-genai>=1.56.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "scikit-learn>=1.8.0",
    "sqlmodel>=0.0.29",
    "uvicorn>=0.40.0",
    "yfinance>=1.0",
]
